+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Get-IdoitConstant
   - Get-IdoitObject, Get-IdoitObjectType, Get-IdoitObjectTypeCategory, Get-IdoitObjectTypeGroup
   - Get-IdoitVersion
+  - Start-IdoitApiTrace, Stop-IdoitApiTrace
 
 ### Changed